fix: Member Entity ProfilUrl 제거
