+++ src/main/java/com/justdo/plug/auth/domain/member/dto/MemberInfoRequest.java
@@ -1,4 +1,4 @@
-package com.justdo.plug.auth.domain.member.dto.request;
+package com.justdo.plug.auth.domain.member.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AccessLevel;
@@ -9,17 +9,15 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
 public class MemberInfoRequest {
-    @Schema(description = "유저 이메일",example = "ht0729@gachon.ac.kr")
+
+    @Schema(description = "유저 이메일", example = "ht0729@gachon.ac.kr")
     private String email;
-    @Schema(description = "유저 닉네임",example = "정성실")
+    @Schema(description = "유저 닉네임", example = "정성실")
     private String nickname;
-    @Schema(description = "유저 프로필 url",example = "https://xxx.xxx.asdf")
-    private String profileUrl;
 
     @Builder
     public MemberInfoRequest(String email, String nickname, String profileUrl) {
         this.email = email;
         this.nickname = nickname;
-        this.profileUrl = profileUrl;
     }
 }
